# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"_36_Enums.Role": {
				"type": "string",
				"enum": [
					"ADMIN",
					"OPERATIONS",
					"SALES_MANAGER"
				]
			},
			"_36_Enums.UserStatus": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"INACTIVE",
					"DELETED"
				]
			},
			"DefaultSelection_Prisma._36_UserPayload_": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.UserStatus"
					},
					"role": {
						"$ref": "#/components/schemas/_36_Enums.Role"
					},
					"passwordHash": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"status",
					"role",
					"passwordHash",
					"username",
					"id"
				],
				"type": "object"
			},
			"User": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_UserPayload_",
				"description": "Model User"
			},
			"Role": {
				"$ref": "#/components/schemas/_36_Enums.Role"
			},
			"UserCreationParams": {
				"properties": {
					"username": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"_36_Enums.EntityStatus": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"DELETED",
					"SUSPENDED",
					"BLACKLISTED"
				]
			},
...
			"EntityStatus": {
				"$ref": "#/components/schemas/_36_Enums.EntityStatus"
			},
...
			"_36_Enums.POStatus": {
				"type": "string",
				"enum": [
					"DELETED",
					"DRAFT",
					"PENDING_L1",
					"APPROVED_L1",
					"REJECTED_L1",
					"ORDER_PLACED",
					"IN_PRODUCTION",
					"QUALITY_INSPECTION",
					"READY_TO_SHIP",
					"SHIPPED",
					"IN_TRANSIT",
					"PORT_CLEARANCE",
					"DELIVERED",
					"RETURNED",
					"CLOSED",
					"CANCELLED"
				]
			},
			"DefaultSelection_Prisma._36_PurchaseOrderPayload_": {
				"properties": {
					"divisionId": {
						"type": "number",
						"format": "double"
					},
					"remarks": {
						"type": "string"
					},
					"poDate": {
						"type": "string",
						"format": "date-time"
					},
					"supplierId": {
						"type": "number",
						"format": "double"
					},
					"poNumber": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.POStatus"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"divisionId",
					"remarks",
					"poDate",
					"supplierId",
					"poNumber",
					"updatedAt",
					"createdAt",
					"status",
					"id"
				],
				"type": "object"
			},
			"PurchaseOrder": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_PurchaseOrderPayload_",
				"description": "Model PurchaseOrder"
			},
			"PaginatedResult_PurchaseOrder_": {
				"description": "Shared pagination types used across all services.",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/PurchaseOrder"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePOParams": {
				"properties": {
					"supplierId": {
						"type": "number",
						"format": "double"
					},
					"divisionId": {
						"type": "number",
						"format": "double"
					},
					"remarks": {
						"type": "string"
					},
					"poDate": {
						"type": "string",
						"format": "date-time"
					},
					"items": {
						"items": {
							"properties": {
								"expectedDeliveryDate": {
									"type": "string",
									"format": "date-time"
								},
								"remarks": {
									"type": "string"
								},
								"totalPrice": {
									"type": "number",
									"format": "double"
								},
								"unitPrice": {
									"type": "number",
									"format": "double"
								},
								"quantity": {
									"type": "number",
									"format": "double"
								},
								"sku": {
									"type": "string"
								},
								"productName": {
									"type": "string"
								},
								"productId": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"totalPrice",
								"unitPrice",
								"quantity"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"supplierId",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreatePOParams-and-_status-POStatus__": {
				"properties": {
					"supplierId": {
						"type": "number",
						"format": "double"
					},
					"divisionId": {
						"type": "number",
						"format": "double"
					},
					"remarks": {
						"type": "string"
					},
					"poDate": {
						"type": "string",
						"format": "date-time"
					},
					"items": {
						"items": {
							"properties": {
								"expectedDeliveryDate": {
									"type": "string",
									"format": "date-time"
								},
								"remarks": {
									"type": "string"
								},
								"totalPrice": {
									"type": "number",
									"format": "double"
								},
								"unitPrice": {
									"type": "number",
									"format": "double"
								},
								"quantity": {
									"type": "number",
									"format": "double"
								},
								"sku": {
									"type": "string"
								},
								"productName": {
									"type": "string"
								},
								"productId": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"totalPrice",
								"unitPrice",
								"quantity"
							],
							"type": "object"
						},
						"type": "array"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.POStatus"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"DefaultSelection_Prisma._36_DivisionPayload_": {
				"properties": {
					"contactPerson": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.EntityStatus"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"contactPerson",
					"updatedAt",
					"createdAt",
					"status",
					"id",
					"name"
				],
				"type": "object"
			},
			"Division": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_DivisionPayload_",
				"description": "Model Division"
			},
			"PaginatedResult_Division_": {
				"description": "Shared pagination types used across all services.",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Division"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DivisionCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"contactPerson": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/EntityStatus"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_DivisionCreationParams_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"contactPerson": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.EntityStatus"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"Pick_User.Exclude_keyofUser.passwordHash__": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/_36_Enums.Role"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.UserStatus"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"role",
					"status",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.passwordHash_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.passwordHash__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"AuthResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/Omit_User.passwordHash_"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginParams": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterParams": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"username",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuditLogResponse": {
				"description": "DTO for API responses — avoids Prisma's `JsonValue` type that tsoa can't resolve",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"entityType": {
						"type": "string"
					},
					"entityId": {
						"type": "number",
						"format": "double"
					},
					"action": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"previousData": {},
					"newData": {},
					"metadata": {},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"entityType",
					"entityId",
					"action",
					"userId",
					"username",
					"previousData",
					"newData",
					"metadata",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResult_AuditLogResponse_": {
				"description": "Shared pagination types used across all services.",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/AuditLogResponse"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "kdm-procurex-backend",
		"version": "1.0.0",
		"description": "KDM ProcureX Backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			}
		},
...
		"/users/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/orders/paginated": {
			"get": {
				"operationId": "GetOrdersPaginated",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_PurchaseOrder_"
								}
							}
						}
					}
				},
				"tags": [
					"Purchase Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders": {
			"get": {
				"operationId": "GetAllOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PurchaseOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Purchase Orders"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PurchaseOrder"
								}
							}
						}
					}
				},
				"tags": [
					"Purchase Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePOParams"
							}
						}
					}
				}
			}
		},
		"/orders/{id}": {
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/PurchaseOrder"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Purchase Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PurchaseOrder"
								}
							}
						}
					}
				},
				"tags": [
					"Purchase Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_CreatePOParams-and-_status-POStatus__"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOrder",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Purchase Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/divisions/paginated": {
			"get": {
				"operationId": "GetDivisionsPaginated",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_Division_"
								}
							}
						}
					}
				},
				"tags": [
					"Division"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/divisions": {
			"get": {
				"operationId": "GetDivisions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Division"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Division"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateDivision",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Division"
								}
							}
						}
					}
				},
				"tags": [
					"Division"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DivisionCreationParams"
							}
						}
					}
				}
			}
		},
		"/divisions/{id}": {
			"get": {
				"operationId": "GetDivision",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Division"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Division"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateDivision",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Division"
								}
							}
						}
					}
				},
				"tags": [
					"Division"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_DivisionCreationParams_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteDivision",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Division"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginParams"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterParams"
							}
						}
					}
				}
			}
		},
		"/auth/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": []
			}
		},
		"/auth/users/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/audit/paginated": {
			"get": {
				"operationId": "GetAuditLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_AuditLogResponse_"
								}
							}
						}
					}
				},
				"description": "Get paginated audit logs (admin only)",
				"tags": [
					"Audit Logs"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "entityType",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "action",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "entityId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/audit/{entityType}/{entityId}": {
			"get": {
				"operationId": "GetEntityHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AuditLogResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get audit history for a specific entity",
				"tags": [
					"Audit Logs"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityType",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/v1"
		}
	]
}